requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# LLM Providers (install at least one)
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
# langchain-google-genai>=2.0.0  # Uncomment for Gemini

# Vector Store
faiss-cpu>=1.8.0

# Document Loaders
pypdf>=4.0.0
beautifulsoup4>=4.12.0

# BM25 Retrieval
rank-bm25>=0.2.2

# Web UI
streamlit>=1.40.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Utilities
tiktoken>=0.7.0
